fix: Transpile node for different js build systems
We need to transpile nodeJS code as well as the browserify in order
to allow build systems that bundle without transpiling to work.

While this is a step in a direction we didn't really want to take
(transpiling everything, even if tiny steps) it will make life alot
easier and has proven to be more popular with contributors.